+++ python/pyproject.toml
@@ -14,6 +14,7 @@ databases = "^0.9.0"
 aiosqlite = "^0.20.0"
 pytest = "^8.3.3"
 httpx = "^0.27.2"
+pytest-asyncio = "^0.25.1"
 
 
 [build-system]
